+++ CHANGELOG.md
@@ -7,6 +7,8 @@ New features:
 - Added Claude Opus 4.5 support via the Anthropic AI backend.
 - Added Kimi-k2, Minimax-m2, GLM-4.6, and Gemini3-Pro-Preview via the Ollama AI backend.  Updated GLM and Gemini3 3 via the
   Z.ai and Google backends to differentiate the two ways to access the same things.
+- Added a series of new features to the system AI tool to allow AIs to interact with conversation tabs.
+- Added a series of new features to the system AI tool to allow AIs to interact with the mindspace log.
 
 ## v0.31 (2025-11-25)